+++ lib/bs_trees/autoload.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
-module BSTrees # :nodoc:
+module BSTrees
+  autoload :Base, 'bs_trees/base'
   autoload :BST, 'bs_trees/bst'
   autoload :AVL, 'bs_trees/avl'
 end