+++ lib/bs_trees/helpers.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'bs_trees/helpers/freezeable'
+require 'bs_trees/helpers/print_methods'
 
 module BSTrees
   module Helpers; end